Fix lifecycle handling on iOS 13+
Fixes #1012.
